+++ Spatial.Core/src/render/SkyboxResources.cpp
@@ -7,7 +7,7 @@
 namespace spatial::render
 {
 
-Texture createKtxTexture(filament::Engine& engine, const uint8_t* data, uint32_t size)
+Texture createKtxTexture(filament::Engine& engine, const uint8_t* data, size_t size)
 {
 	auto ktxBundle = new image::KtxBundle(data, size);
 	return Texture{engine, image::ktx::createTexture(&engine, ktxBundle, true)};
